TVB-2512 rename tvb.core.entities.file.datatypes.time_series into tvb.core.entities.file.datatypes.time_series_h5 for consistency
