+++ tvb/adapters/analyzers/pca_adapter.py
@@ -45,7 +45,7 @@ from tvb.datatypes.mode_decompositions import PrincipalComponents
 from tvb.core.entities.filters.chain import FilterChain
 from tvb.basic.logger.builder import get_logger
 from tvb.core.entities.file.datatypes.mode_decompositions_h5 import PrincipalComponentsH5
-from tvb.core.entities.file.datatypes.time_series import TimeSeriesH5
+from tvb.core.entities.file.datatypes.time_series_h5 import TimeSeriesH5
 from tvb.core.entities.model.datatypes.mode_decompositions import PrincipalComponentsIndex
 from tvb.core.entities.model.datatypes.time_series import TimeSeriesIndex
 from tvb.core.neotraits._forms import DataTypeSelectField
