+++ tvb/adapters/uploaders/gifti_timeseries_importer.py
@@ -38,7 +38,7 @@ from tvb.core.adapters.abcuploader import ABCUploader, ABCUploaderForm
 from tvb.core.adapters.exceptions import LaunchException, ParseException
 from tvb.adapters.uploaders.gifti.parser import GIFTIParser
 from tvb.basic.logger.builder import get_logger
-from tvb.core.entities.file.datatypes.time_series import TimeSeriesSurfaceH5
+from tvb.core.entities.file.datatypes.time_series_h5 import TimeSeriesSurfaceH5
 from tvb.core.entities.model.datatypes.surface import SurfaceIndex
 from tvb.core.entities.model.datatypes.time_series import TimeSeriesSurfaceIndex
 from tvb.core.neotraits._forms import UploadField, DataTypeSelectField
