+++ tvb/tests/framework/core/entities/file/datatypes/time_series_test.py
@@ -1,5 +1,35 @@
+# -*- coding: utf-8 -*-
+#
+#
+# TheVirtualBrain-Framework Package. This package holds all Data Management, and
+# Web-UI helpful to run brain-simulations. To use it, you also need do download
+# TheVirtualBrain-Scientific Package (for simulators). See content of the
+# documentation-folder for more details. See also http://www.thevirtualbrain.org
+#
+# (c) 2012-2017, Baycrest Centre for Geriatric Care ("Baycrest") and others
+#
+# This program is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software Foundation,
+# either version 3 of the License, or (at your option) any later version.
+# This program is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+# PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+# You should have received a copy of the GNU General Public License along with this
+# program.  If not, see <http://www.gnu.org/licenses/>.
+#
+#
+#   CITATION:
+# When using The Virtual Brain for scientific publications, please cite it as follows:
+#
+#   Paula Sanz Leon, Stuart A. Knock, M. Marmaduke Woodman, Lia Domide,
+#   Jochen Mersmann, Anthony R. McIntosh, Viktor Jirsa (2013)
+#       The Virtual Brain: a simulator of primate brain network dynamics.
+#   Frontiers in Neuroinformatics (7:10. doi: 10.3389/fninf.2013.00010)
+#
+#
+
 import numpy
-from tvb.core.entities.file.datatypes.time_series import TimeSeriesH5
+from tvb.core.entities.file.datatypes.time_series_h5 import TimeSeriesH5
 from tvb.datatypes.time_series import TimeSeries
 
 
@@ -12,8 +42,7 @@ def make_harmonic_ts():
         labels_ordering=('time', 'statevar', 'space'),
         labels_dimensions={'statevar': ['position', 'speed']},
         start_time=0.0,
-        sample_period=0.5,
-        sample_rate=2.0,
+        sample_period=0.5
     )
